refactor: move board-session & grid span logic into shared core/domain (issue #130)
