+++ composeApp/src/commonTest/kotlin/io/github/jdreioe/wingmate/ui/BoardEditorGraphTest.kt
@@ -1,14 +1,24 @@
 package io.github.jdreioe.wingmate.ui
 
 import io.github.jdreioe.wingmate.domain.obf.BoardSetGraph
+import io.github.jdreioe.wingmate.domain.obf.BoardSettingsOverrides
 import io.github.jdreioe.wingmate.domain.obf.ObfBoard
 import io.github.jdreioe.wingmate.domain.obf.ObfBoardSet
 import io.github.jdreioe.wingmate.domain.obf.ObfButton
 import io.github.jdreioe.wingmate.domain.obf.ObfGrid
+import io.github.jdreioe.wingmate.domain.obf.ObfImage
 import io.github.jdreioe.wingmate.domain.obf.ObfLoadBoard
-import io.github.jdreioe.wingmate.domain.obf.BoardSettingsOverrides
+import io.github.jdreioe.wingmate.domain.obf.ObfSound
 import io.github.jdreioe.wingmate.domain.obf.pageSettingsOverrides
 import io.github.jdreioe.wingmate.domain.obf.withPageSettingsOverrides
+import io.github.jdreioe.wingmate.domain.obf.updateDraftCell
+import io.github.jdreioe.wingmate.domain.obf.clearDraftCell
+import io.github.jdreioe.wingmate.domain.obf.renameDraftBoardSet
+import io.github.jdreioe.wingmate.domain.obf.renameDraftBoard
+import io.github.jdreioe.wingmate.domain.obf.resizeDraftBoard
+import io.github.jdreioe.wingmate.domain.obf.moveDraftField
+import io.github.jdreioe.wingmate.domain.obf.resizeDraftField
+import io.github.jdreioe.wingmate.domain.obf.withHomeFieldsBottomLeft
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNotEquals
