+++ core/domain/src/commonTest/kotlin/io/github/jdreioe/wingmate/domain/obf/BoardCellTapTest.kt
@@ -1,7 +1,5 @@
-package io.github.jdreioe.wingmate.ui
+package io.github.jdreioe.wingmate.domain.obf
 
-import io.github.jdreioe.wingmate.domain.obf.ObfButton
-import io.github.jdreioe.wingmate.domain.obf.ObfGrid
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertIs
