--- linuxApp/compile_commands.json
@@ -1,1 +1,0 @@
-build/compile_commands.json
