--- linuxApp/src/components/qmldir
@@ -1,11 +1,0 @@
-module components
-singleton Theme 1.0 Theme.qml
-ModernCard 1.0 ModernCard.qml
-Sidebar 1.0 Sidebar.qml
-ModernButton 1.0 ModernButton.qml
-PhraseButton 1.0 PhraseButton.qml
-Chip 1.0 Chip.qml
-BottomBar 1.0 BottomBar.qml
-SpeechControlsSidebar 1.0 SpeechControlsSidebar.qml
-PredictionBar 1.0 PredictionBar.qml
-OnScreenKeyboard 1.0 OnScreenKeyboard.qml
