--- linuxApp/src/pages/qmldir
@@ -1,5 +1,0 @@
-module pages
-SettingsPage 1.0 SettingsPage.qml
-PronunciationDictionaryPage 1.0 PronunciationDictionaryPage.qml
-AddPronunciationEntryPage 1.0 AddPronunciationEntryPage.qml
-WelcomeWizard 1.0 WelcomeWizard.qml
